+++ R/periodic.R
@@ -11,10 +11,10 @@
 #'
 #' An object of subclass `periodic_df` or `periodic_v`.
 #'
-#' @details
-#'
-#' This is a generic function and ggperiodic provides methods for vectors, data frames
-#' and ggplot2 Layers (geoms and stats).
+#' If `object` is of class `data.table`, then it will modify the object by
+#' reference. To modify this behaviour, use
+#' `options(ggperiodic.data.table.copy = TRUE)`. `setperiodic()` will modify a
+#' `data.table` by reference bypassing the global option.
 #'
 #' @examples
 #' library(ggplot2)
@@ -57,12 +57,17 @@ periodic.default <- function(object, period, ...) {
 #' @export
 #' @rdname periodic
 #' @method periodic data.frame
+#' @importFrom data.table copy setattr
 periodic.data.frame <- function(object, ...) {
   cols <- as.list(substitute(list(...))[-1])
 
   if(length(cols) == 0) {
     warning("No columns defined. Returning unchanged data.")
     return(object)
+  }
+
+  if (.should.copy(object)) {
+    object <- data.table::copy(object)
   }
 
   iscol <- names(cols) %in% colnames(object)
@@ -91,8 +96,11 @@ periodic.data.frame <- function(object, ...) {
     if (r[1] < period[1] | r[2] > period[2]) {
       bad.cols <- c(bad.cols, names(cols)[i])
     } else {
-      attr(object[[names(cols)[i]]], "period") <- period
-      class(object[[names(cols)[i]]]) <- c("sticky", class(object[[names(cols)[i]]]))
+      data.table::setattr(object[[names(cols)[i]]], "period", period)
+      if (!inherits(object[[names(cols)[i]]], "sticky")) {
+        data.table::setattr(object[[names(cols)[i]]], "class", c("sticky", class(object[[names(cols)[i]]])))
+        # class(object[[names(cols)[i]]]) <- c("sticky", class(object[[names(cols)[i]]]))
+      }
     }
   }
   if (length(bad.cols) == length(cols)) {
@@ -105,9 +113,15 @@ periodic.data.frame <- function(object, ...) {
   }
 
   if (!inherits(object, "periodic_df")) {
-    class(object) <- c("periodic_df", class(object))
+    data.table::setattr(object, "class", c("periodic_df", class(object)))
+    # class(object) <-  c("periodic_df", class(object))
   }
-  object
+
+  if (.should.copy(object)) {
+    return(object)
+  } else {
+    return(invisible(object))
+  }
 }
 
 
@@ -122,3 +136,11 @@ print.periodic_df <- function(x, ...) {
   }
 }
 
+
+#' @export
+#' @rdname periodic
+setperiodic <- function(object, ...) {
+  old <- .set.copy(FALSE)
+  on.exit(.set.copy(old))
+  periodic(object, ...)
+}