+++ R/utils.R
@@ -13,4 +13,19 @@ join <- function(x, y) {
   j <- c(x, y)
   j[unique(names(j))]
 }
+
+
+.set.copy <- function(value) {
+  old <- .get.copy()
+  options(ggperiodic.data.table.copy = value)
+  old
+}
+
+.get.copy <- function() {
+  getOption("ggperiodic.data.table.copy", default = TRUE)
+}
+
+.should.copy <- function(object) {
+  !inherits(object, "data.table") | .get.copy()
+}
 # nocov end