+++ NEWS.md
@@ -1,3 +1,9 @@
+# ggperiodic 0.1.9000
+
+* By default `periodic()` and `unperiodic()` will modify by reference if 
+`object` is a `data.table`. To modify this behaviour, use 
+`options(ggperiodic.data.table.copy = TRUE)`.
+
 # ggperiodic 0.1.0
 
 * Implemented `dplyr` methods.